requirements.txt
alembic==1.6.5
amqp==5.0.6
aniso8601==9.0.1
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.1.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
dropbox==11.10.0
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
Mako==1.1.4
MarkupSafe==2.0.1
ply==3.11
prompt-toolkit==3.0.18
protobuf==3.17.3
pycparser==2.20
pydotenv==0.0.7
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
redispy==3.0.0
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.18
stone==3.2.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
